+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 # Changelog
-## v0.3.0 - Revision 48
+## v0.3.0 
+### Revision 50
+Fix stubs for console logging, add OpenGL calls.
+### Revision 49
+Add stubs for console logging.
+### Revision 48
 New basis for the game engine. Starting this off from scratch.